// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPERADMIN: 'SUPERADMIN',
  ADMIN: 'ADMIN',
  SECURITYPERSON: 'SECURITYPERSON',
  USER: 'USER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AccountStatus = {
  ACTIVE: 'ACTIVE',
  DEACTIVATED: 'DEACTIVATED',
  BANNED: 'BANNED',
  DELETIONPENDING: 'DELETIONPENDING',
  DELETED: 'DELETED'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const BloodGroup = {
  A_POSITIVE: 'A_POSITIVE',
  A_NEGATIVE: 'A_NEGATIVE',
  B_POSITIVE: 'B_POSITIVE',
  B_NEGATIVE: 'B_NEGATIVE',
  AB_POSITIVE: 'AB_POSITIVE',
  AB_NEGATIVE: 'AB_NEGATIVE',
  O_POSITIVE: 'O_POSITIVE',
  O_NEGATIVE: 'O_NEGATIVE'
} as const

export type BloodGroup = (typeof BloodGroup)[keyof typeof BloodGroup]


export const PoliceRank = {
  IGP: 'IGP',
  ADDITIONAL_IGP: 'ADDITIONAL_IGP',
  DIG: 'DIG',
  ADDITIONAL_DIG: 'ADDITIONAL_DIG',
  POLICE_COMMISSIONER: 'POLICE_COMMISSIONER',
  ADDITIONAL_POLICE_COMMISSIONER: 'ADDITIONAL_POLICE_COMMISSIONER',
  SP: 'SP',
  ADDITIONAL_SP: 'ADDITIONAL_SP',
  SENIOR_ASP: 'SENIOR_ASP',
  ASP: 'ASP',
  INSPECTOR: 'INSPECTOR',
  SUB_INSPECTOR: 'SUB_INSPECTOR',
  SERGEANT: 'SERGEANT',
  ASSISTANT_SUB_INSPECTOR: 'ASSISTANT_SUB_INSPECTOR',
  NAYEK: 'NAYEK',
  CONSTABLE: 'CONSTABLE'
} as const

export type PoliceRank = (typeof PoliceRank)[keyof typeof PoliceRank]